[Driver] Handle cases where swiftdeps could not be loaded even after a build.
If a build fails in the middle, we try to determine which other files need
to be rebuilt. However, we may not be able to do that as precisely if the
dependency graph itself is incomplete. In this case, just be conservative
and assume we need to rebuild everything. We may want to revisit this in
the future with a more-aggressive-but-still-safe bound.

This was manifesting itself as an assertion failure, trying to pull
information from the graph that wasn't there.

rdar://problem/19640006

Swift SVN r24823