Consider @objc selectors when matching witnesses to protocol requirements.
When we match a witness to a requirement in a protocol, we do so based
on the Swift name (which is correct). When the requirement is @objc
(because it is in an @objc protocol), also check that the Objective-C
selectors of the witness match those of the requirement.

Fixes the majority of rdar://problem/18383574.

Swift SVN r24829