Allow @IBDesignable on class extensions as well.
Per discussion with the IB team, a class can retroactively be marked as
designable via an extension (or if not retroactively, at least from elsewhere
in the module). This matches the documentation for Objective-C.

(The attribute still has no semantics in Swift itself. The only restriction
is that it must appear on a class or an extension of a class.)

rdar://problem/19654163

Swift SVN r24837