Sema: Allow availability attributes on enum elements.
With this commit, availability attributes can now be placed on enum elements. We
emit an error (protected by the -enable-experimental-availability-checking flag) for
references to potentially unavailable enum cases. We do *not* emit an error when the
programmer pattern matches (i.e., with switch) on a potentially unavailable case.

This commit does not import availability attributes from clang. I will add support for
this in a later commit.

Swift SVN r24929