[stdlib] Don't leak CF dependency from core stdlib
Carefully track the functions that inject these dependencies, so all
functions that use them can be protected with

  @inline(never) @semantics("stdlib_binary_only")

To do this, we introduce a naming convention: the _stdlib_binary_ prefix
distinguishes functions dependent on external frameworks.  All callers
of such functions must adopt either the naming convention, or the
attributes above.

This change should unbreak the build.

Swift SVN r25020