Runtime: Remove lazy symbol lookup hacks for ObjC mirrors.
DaveA fixed things so the mirror implementations all live in the standard library, so we don't need to lazily resolve the necessary symbols in the runtime anymore. While we're here, use __asm__ to hide the mangled Swift symbols a bit in the code that needs to use them.

Swift SVN r25057