stdlib: fast enumeration: stop relying on stack promotion optimization for correctness
Fast enumeration clients rely on pointer identity of the fast
enumeration state struct.  We were using 'withUnsafeMutablePointers' to
get a pointer to the state stored in a property, but we were relying on
stack promotion optimization to access the stored property itself rather
than the writeback buffer.

rdar://19165256

Swift SVN r25146