[ClangImporter] Delay creating the body for struct zero-initializers.
This way, we don't have issues with the verifier bringing in new types and
then those new types having members with un-type-checked bodies. Also,
delaying this is probably future goodness anyway, when we can skip the work
entirely if it's not needed.

I can't come up with a test case that currently fails besides the project in
rdar://problem/19778991, but all existing tests pass.

Swift SVN r25166