Add a simple utility for determining the lifetime of a value.
This refactors some code out of AllocBoxToStack that computes the
lifetime of a value in the strictess sense, limiting the lifetime to
that value and not anything derived from that value (whether by casting,
projection, etc.).

In the short term this will be used to fix a very rarely hit
optimization in AllocBoxToStack.

Longer term I will replace the other similar code in AllocBoxToStack to
use this instead.

Swift SVN r25176