Fix issue with where we release boxes in AllocBoxToStack.
When we specialize a partial_apply to remove boxes that were arguments
to the partial_apply, we release the box explicitly (since it will no
longer get released as part of releasing the partial_apply).

We were putting these releases before the partial_apply, only worked
properly in cases where those releases were not the final release of the
box.

With this commit we now release the box at each point where the
partial_apply goes dead.

Fixes rdar://problem/17892969.

Swift SVN r25177