CopyForwarding pass: make destroy hoisting conservative if the addess is also loaded from.
We never run into this case in the stdlib/PerfTestSuite. So for Xcode 6.4, be conservative.
Longer term, I was planning to follow def-use chains through projections.
When we do this, we can also follow projects of loaded values to find any retain points.
That would give us a proper indication of the lifetime of referenced objects.

Fixes <rdar://problem/19724405> [NSObject] as! [NSWindow] sends messages to deallocated __NSArrayI

Swift SVN r25188