fix <rdar://problem/19772570> Parsing of as and ?? regressed
In r22669, we expanded our parsing of T?? by using operator splitting.  This works
great when the ?? is lexed as a postfix operator, but does the wrong thing when the ??
is lexed as an infix  and there is whitespace around it.  For that, we don't want to 
process it as a type.  This was reported in a number of places including on stack
overflow:
http://stackoverflow.com/questions/28430496/nil-coalescing-to-provide-default-values-in-swift-1-2



Swift SVN r25228