fix <rdar://problem/19783405> @autoclosure parameters 'escape' from compiler-generated default initilizers
This is a "noescape leak" caused by implicit constructor synthesis, and not rejecting the
invalid code that is synthesized.  There isn't a good way to synthesize an implicit memberwise
constructor for a structure with an @autoclosure property, we just surpress it.

It would be great to just disallow autoclosure on properties entirely.  The only really popular
usecase I know of is for breaking cycles in enums, maybe we should just limit the autoclosure attribute
to enums specifically.



Swift SVN r25231