[sil-combine] Correctness fix for checked_cast_addr_br folding.
Don't fold the negative case if the type is internal and we are not doing whole-module-optimization, because an extension may define a conformance in a different file.

Fold the negative case for private types. It is safe, because conformances could be defined only in the current file and we should have seen them.

rdar://19852884

Swift SVN r25331