[stdlib] Downcasted arrays stay native
We used to handle deferred type-checking by treating down-casted native
array buffers as NSArrays (which they are, but we know more).  Instead,
we now save a bit that indicates deferred type-checking is needed and
remember that the buffer is native, which saves dispatching through
objc_MsgSend.

Fixes <rdar://problem/19302286> down-casted Arrays are inefficient

Swift SVN r25472