[callgraph] Add a map from Apply -> CallGraphEdge.
This makes it easy to use the callgraph to determine what the Apply
actually calls instead of attempting to ascertain it in an adhoc
manner. This allows for the call graph analysis to be used as the one
point of truth for determing what an apply inst can call.

Swift SVN r25516